Error fixes, dunno about Socket.h fix, but i think it's ok. Btw, _socket.get_io_service() function is deprecated, check line below.

+++ Game/ArkaWar.h
@@ -115,8 +115,8 @@ static const struct ArkaWarZoneBattleData
 	{WORLD_ACHERON_BATTLE, 427, "Alkmar"},
 	{WORLD_ACHERON_BATTLE, 431, "Ubaid"},
 	{WORLD_DEVENTER_BATTLE, 439, "Debenter"},
-	{-1, -1, ""},
-	{-1, -1, ""}
+	{(uint16) -1, (uint16) -1, ""},
+	{(uint16) -1, (uint16) -1, ""}
 };
 
 struct ArkaWarObeliskAuraBasicData
